Improve consistency wrt use of colons
The logic is this: If the key and value are on the same line, then
follow the value by a colon.  Otherwise, if the value is a list, then
each element is displayed on its own line and the key is not followed
by a colon because there is no need for some additional separator.